Fixed the ExpandingFactWidget so it expands keeping it's aspect ratio.

+++ V3/src/widgets/expandingFactWidget.h
@@ -29,7 +29,7 @@ private:
     QString title;
     QPixmap image, scaledImage;
     bool expanded;
-    int currentHeight;
+    double currentHeight;
 
     QWidget *headWidget;
     int headWidgetHeight;
